REFACTOR :: components/ 폴더 아래에 index.ts 파일 위치 변경
