+++ apps/user/src/components/faq/index.ts
@@ -1,0 +1,2 @@
+export { default as FaqItem } from './FaqItem/FaqItem';
+export { default as CategoryFilter } from './CategoryFilter/CategoryFilter';
