--- apps/user/src/components/index.ts
@@ -1,14 +1,0 @@
-export { default as FaqItem } from './faq/FaqItem/FaqItem';
-export { default as CategoryFilter } from './faq/CategoryFilter/CategoryFilter';
-
-export { default as ProgressBar } from './form/ProgressBar/ProgressBar';
-
-export { default as Dday } from './main/Dday/Dday';
-export { default as Faq } from './main/Faq/Faq';
-export { default as Notice } from './main/Notice/Notice';
-export { default as Schedule } from './main/Schedule/Schedule';
-
-export { default as NoticeHeader } from './notice/NoticeHeader/NoticeHeader';
-export { default as NoticeItem } from './notice/NoticeItem/NoticeItem';
-
-export { default as Terms } from './signup/Terms/Terms';
