+++ apps/user/src/components/main/index.ts
@@ -1,0 +1,4 @@
+export { default as Dday } from './Dday/Dday';
+export { default as Faq } from './Faq/Faq';
+export { default as Notice } from './Notice/Notice';
+export { default as Schedule } from './Schedule/Schedule';
