+++ apps/user/src/components/notice/index.ts
@@ -1,0 +1,2 @@
+export { default as NoticeHeader } from './NoticeHeader/NoticeHeader';
+export { default as NoticeItem } from './NoticeItem/NoticeItem';
