+++ apps/user/src/components/signup/index.ts
@@ -1,0 +1,1 @@
+export { default as Terms } from './Terms/Terms';
